Replace C variable types to specific sized vars.
Replaced all references to
1) unsigned long
2) unsigned int
3) unsigned char
4) unsigned long long

to the appropriate constant and explicit length
1) uint32_t
2) uint16_t
3) uint8_t
4) uint64_t

Since we have to interface to dnvme in these explicit terms it makes more sense
to explicitly state the size and types we are needing throughout all of the
code. This hopefully will avoid the issue of whether or not a unsigned int is
16 or 32 bits.

+++ GrpInformative/dumpPciAddrSpace_r10b.h
@@ -20,7 +20,7 @@ protected:
 
 private:
     void WriteToFile(int fd, const PciSpcType regMetrics,
-        unsigned long long value);
+        uint64_t value);
 };
